898: Removing ParseCertificate script from Test Trusted Directory get SSA route
Moving logic from ParseCertificate script into JwkmsBuildSSA script, as this is the only place that it is needed.

Updating route to use the new filter: AddCertificateToAttributesContextFilter which adds the cert to the context.

Removing ParseCertificate.groovy as it is no longer used.

https://github.com/SecureApiGateway/SecureApiGateway/issues/898
